+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/rings/RingOfMight.java
@@ -110,7 +110,7 @@ public class RingOfMight extends Ring {
 	}
 	
 	public static int strengthBonus( Char target ){
-		return getBonus( target, Might.class );
+		return (int) getBonus( target, Might.class );
 	}
 	
 	public static float HTMultiplier( Char target ){
